Improve performance of resolvePageTree
