+++ src/lib/pages/resolvePageTree.ts
@@ -1,4 +1,4 @@
-import type { Prisma } from '@prisma/client';
+import type { Page, Prisma } from '@prisma/client';
 
 import type { OptionalPrismaTransaction } from '../../prisma-client';
 import { prisma } from '../../prisma-client';
@@ -9,19 +9,26 @@ import type {
   PageTreeResolveInput,
   TargetPageTree,
   TargetPageTreeWithFlatChildren,
-  PageWithPermissions
+  PageWithPermissions,
+  PageMetaWithPermissions
 } from './interfaces';
 import { flattenTree, mapTargetPageTree } from './mapPageTree';
 
 function generatePagesQuery({
-  spaceId,
   includeDeletedPages,
-  fullPage
+  fullPage,
+  pageIds,
+  spaceId
 }: {
-  spaceId: string;
   includeDeletedPages?: boolean;
   fullPage?: boolean;
+  pageIds?: string[];
+  spaceId?: string;
 }) {
+  if (!pageIds && !spaceId) {
+    throw new InvalidInputError(`1 of spaceId or pageIds is required`);
+  }
+
   const pageQueryContent: Partial<Prisma.PageFindManyArgs> = fullPage
     ? {
         include: {
@@ -53,6 +60,9 @@ function generatePagesQuery({
   return {
     where: {
       spaceId,
+      id: {
+        in: pageIds
+      },
       // Soft deleted pages have a value for deletedAt. Active pages are null
       deletedAt: includeDeletedPages ? undefined : null
     },
@@ -72,8 +82,7 @@ function generatePagesQuery({
 export async function resolvePageTree({
   pageId,
   flattenChildren,
-  fullPage,
-  pageNodes
+  fullPage
 }: PageTreeResolveInput & {
   flattenChildren?: undefined | false;
   fullPage?: false | undefined;
@@ -81,8 +90,7 @@ export async function resolvePageTree({
 export async function resolvePageTree({
   pageId,
   flattenChildren,
-  fullPage,
-  pageNodes
+  fullPage
 }: PageTreeResolveInput & { flattenChildren: true; fullPage?: false | undefined } & OptionalPrismaTransaction): Promise<
   TargetPageTreeWithFlatChildren<PageNodeWithPermissions>
 >;
@@ -90,16 +98,14 @@ export async function resolvePageTree({
 export async function resolvePageTree({
   pageId,
   flattenChildren,
-  fullPage,
-  pageNodes
+  fullPage
 }: PageTreeResolveInput & { flattenChildren?: undefined | false; fullPage: true } & OptionalPrismaTransaction): Promise<
   TargetPageTree<PageWithPermissions>
 >;
 export async function resolvePageTree({
   pageId,
   flattenChildren,
-  fullPage,
-  pageNodes
+  fullPage
 }: PageTreeResolveInput & { flattenChildren: true; fullPage: true } & OptionalPrismaTransaction): Promise<
   TargetPageTreeWithFlatChildren<PageWithPermissions>
 >;
@@ -108,34 +114,55 @@ export async function resolvePageTree({
   flattenChildren = false,
   includeDeletedPages = false,
   fullPage,
-  pageNodes,
   tx = prisma
 }: PageTreeResolveInput & OptionalPrismaTransaction): Promise<
   TargetPageTree<PageNodeWithPermissions> | TargetPageTreeWithFlatChildren<PageNodeWithPermissions>
 > {
-  const pageWithSpaceIdOnly = pageNodes
-    ? pageNodes.find((node) => node.id === pageId)
-    : await tx.page.findUnique({
-        where: {
-          id: pageId
-        },
-        select: {
-          spaceId: true
-        }
-      });
+  const pageWithSpaceIdOnly = await tx.page.findUnique({
+    where: {
+      id: pageId
+    },
+    select: {
+      spaceId: true
+    }
+  });
 
   if (!pageWithSpaceIdOnly) {
     throw new PageNotFoundError(pageId);
   }
+  const pagesTreeIds = (await tx.$queryRaw`WITH RECURSIVE parents_cte AS (
+    SELECT id, "parentId"
+    FROM "Page"
+    WHERE id = ${pageId}::UUID
+    
+    UNION ALL
+    
+    SELECT p.id, p."parentId"
+    FROM "Page" p
+    INNER JOIN parents_cte ON p.id = parents_cte."parentId"
+  ), children_cte AS (
+    SELECT id, "parentId"
+    FROM "Page"
+    WHERE id = ${pageId}::UUID
+    
+    UNION ALL
+    
+    SELECT p.id, p."parentId"
+    FROM "Page" p
+    INNER JOIN children_cte ON p."parentId" = children_cte.id
+  )
+  SELECT id FROM parents_cte 
+  UNION 
+  SELECT id FROM children_cte;
+`) as Pick<Page, 'id'>[];
 
-  const pagesInSpace = (pageNodes ??
-    (await tx.page.findMany(
-      generatePagesQuery({
-        includeDeletedPages,
-        spaceId: pageWithSpaceIdOnly.spaceId,
-        fullPage
-      })
-    ))) as PageNodeWithPermissions[] | PageWithPermissions[];
+  const pagesInSpace = (await tx.page.findMany(
+    generatePagesQuery({
+      includeDeletedPages,
+      pageIds: pagesTreeIds?.map((p) => p.id) ?? [],
+      fullPage
+    })
+  )) as PageNodeWithPermissions[] | PageWithPermissions[];
   const { parents, targetPage } = mapTargetPageTree({
     items: pagesInSpace,
     targetPageId: pageId,
@@ -160,7 +187,7 @@ export async function resolvePageTree({
     parents,
     targetPage,
     flatChildren: flattenChildren ? flattenTree(targetPage) : undefined
-  };
+  } as any;
 }
 
 export type MultiPageTreeResolveInput<F extends boolean | undefined = boolean> = Pick<
@@ -219,27 +246,65 @@ export async function multiResolvePageTree<F extends boolean | undefined>({
     return {};
   }
 
-  const spaceId = uniqueSpaceIds[0];
+  const pagesTreeIds = (await tx.$queryRaw`WITH RECURSIVE parents_cte AS (
+    SELECT id, "parentId"
+    FROM "Page"
+    WHERE id = ANY(ARRAY[${pageIds}]::UUID[])
+    
+    UNION ALL
+    
+    SELECT p.id, p."parentId"
+    FROM "Page" p
+    INNER JOIN parents_cte ON p.id = parents_cte."parentId"
+  ), children_cte AS (
+    SELECT id, "parentId"
+    FROM "Page"
+    WHERE id = ANY(ARRAY[${pageIds}]::UUID[])
+    
+    UNION ALL
+    
+    SELECT p.id, p."parentId"
+    FROM "Page" p
+    INNER JOIN children_cte ON p."parentId" = children_cte.id
+  )
+  SELECT id FROM parents_cte 
+  UNION 
+  SELECT id FROM children_cte;
+`) as Pick<Page, 'id'>[];
 
   const pagesInSpace = (await tx.page.findMany(
     generatePagesQuery({
       includeDeletedPages,
-      spaceId
+      pageIds: pagesTreeIds.map((p) => p.id),
+      fullPage
     })
   )) as PageNodeWithPermissions[] | PageWithPermissions[];
 
-  const mappedResults = await Promise.all(
-    pageIds.map((id) =>
-      resolvePageTree({
-        pageId: id,
-        flattenChildren: flattenChildren as any,
-        includeDeletedPages,
-        fullPage: fullPage as any,
-        pageNodes: pagesInSpace,
-        tx
-      }).catch(() => null)
-    )
+  const pagemap = (pagesInSpace as PageNodeWithPermissions[]).reduce((acc, val) => {
+    acc[val.id] = val;
+    return acc;
+  }, {} as Record<string, PageNodeWithPermissions>);
+
+  let mappedResults = pageIds.map((id) =>
+    pagemap[id]
+      ? mapTargetPageTree({
+          items: pagesInSpace,
+          targetPageId: id,
+          includeDeletedPages
+        })
+      : null
   );
+
+  if (flattenChildren) {
+    mappedResults = mappedResults.map((tree) => {
+      return tree
+        ? ({
+            ...tree,
+            flatChildren: flattenTree(tree.targetPage)
+          } as TargetPageTreeWithFlatChildren<PageMetaWithPermissions>)
+        : null;
+    });
+  }
 
   return pageIds.reduce((acc, id, index) => {
     acc[id] = mappedResults[index] as any;
